requirements.txt
# Core LLM + Google Gemini
google-genai

# Web scraping
playwright==1.43.0
beautifulsoup4==4.12.3

# Embedding + Semantic Search
sentence-transformers==2.7.0
# Vector DB for versioning and retrieval
chromadb==0.4.24

# Machine Learning for RL
scikit-learn>=1.3.0
numpy>=1.24.0

# Utilities
argparse
python-dotenv
--extra-index-url https://download.pytorch.org/whl/cu128
torch
